+++ front/app/api/graph_data_units/responseTypes/ActiveUsersWidget.ts
@@ -1,0 +1,14 @@
+export interface TimeSeriesResponseRow extends ActiveUsersRow {
+  first_dimension_date_created_date: string;
+}
+
+export interface ActiveUsersRow {
+  count_participant_id: number;
+}
+
+export interface ActiveUsersResponse {
+  data: {
+    type: 'report_builder_data_units';
+    attributes: [TimeSeriesResponseRow[] | [], [ActiveUsersRow] | []];
+  };
+}
