+++ front/app/containers/Admin/reporting/components/ReportBuilder/Widgets/ChartWidgets/ActiveUsersWidget/useActiveUsers/typings.ts
@@ -6,14 +6,6 @@ import {
 
 export type QueryParameters = ProjectId & DatesStrings & Resolution;
 
-export interface TimeSeriesResponseRow extends ActiveUsersRow {
-  first_dimension_date_created_date: string;
-}
-
-export interface ActiveUsersRow {
-  count_participant_id: number;
-}
-
 // Hook return value
 export interface TimeSeriesRow {
   /* Date format: YYYY-MM-DD */
